feat: Add mention functionality to posts
Introduces the ability for users to mention other users in post content using the "@" symbol. This includes a UI component for displaying mention suggestions and parsing mentions to link to user profiles.

The `CreateFeedItemForm` now accepts `allUsers` and manages mention query state, including displaying suggestions. The `PostCard` component is updated to render mentioned users as clickable links, triggering the `onViewProfile` handler. A new `MentionSuggestions` component is created for the suggestion dropdown.